Use floats for edges and distances, clean up A*
- Replaced ints with floats for edge weights and distance computations
to make heuristics etc. more accurate

- Distance between coords is now calculated with Euclidean distance
instead of Manhattan distance. This is more accurate.

- Cleaned up A*: added comments, removed redundant `visited` vec etc.

+++ src/pages/pathfinding.rs
@@ -21,7 +21,7 @@ use yew::prelude::*;
 use yew_hooks::use_title;
 use yew_router::prelude::*;
 
-pub type EdgeType = isize;
+pub type EdgeType = f32;
 
 type PathfindingFunc<V, E> =
     fn(AdjacencyList<V, E>, V, V, PathfindingSteps<V>) -> PathfindingResult<V, E>;
